geotiff tests: consolidate parity tail cluster (#2431)
Fold eight top-level parity test files into three focused siblings next
to parity/test_backend_matrix.py, grouped by concern:

- parity/test_finalization.py: read-finalization parity (dispatch
  validation #2175, eager #2179, lazy #2178; epic #2162).
- parity/test_signature_contract.py: signature / docstring / release-
  contract parity (#1631, #2274, #2389).
- parity/test_reference.py: degenerate-shape backend parity and the
  rasterio / Zarr external-reference round trip (2026-05-11 sweep, #1961).

backend_matrix.py is already 2090 lines; appending all eight files would
push it past 4600. Three siblings keep each file well under the 2500-line
guidance in the issue and group related tests together.

GPU gating moves to the shared requires_gpu marker from
_helpers/markers.py, replacing the per-file _gpu_available helpers. No
assertion changed; the three files collect the same 137 tests as the
eight source files.

CLUSTER_AUDIT_PARITY.md maps every old file::test to its new
file::test_id and is deleted in a final pre-merge commit per epic #2424.

Tests-only -- no source changes.
